Renaming to NBTApi and NBTInjector
For all intents and purposes the libs are now called NBTApi and NBTInjector. In some places it will still be ItemNBTApi(github/bStats/maven).
The NBTApi now needs to be shaded to a different package in order to be used in plugins, else you will get a warning in the console.